Add preload scene with AudioManager
The preload scene is loaded at the start of the
game, and have some objects that are usefull for
all other scenes of the game.

The AudioManager objects store clips and can play,
pause, unpause or stop clips using any script in any
scene.

Finally, every scene now have a 'OnlyForDevelopment'
object, used to fast jump to the Preload scene and
get back to the current scene being developed.